connent to post endpoint
