+++ src/Pages/ProfilePage/ProfilePage.scss
@@ -15,13 +15,13 @@
 
 @keyframes map-out {
   0%{
-    opacity: 0.8;
+    opacity: 0.65;
     scale: .45;
-    transform: translateY(-50vh) translateX(-10vw);
+    transform: translateY(-60vh) translateX(-10vw);
   }
   
   25%{
-    transform: translateY(-70vh)translateX(-15vw);
+    transform: translateY(-90vh)translateX(-15vw);
     z-index: 1000;
   }
   
@@ -32,7 +32,7 @@
     transform:translateY(-10vh) translateX(-12vw);
   }
   100%{
-    opacity: 0.6;
+    opacity: 0.4;
     scale: .2;
     display: none;
   }
@@ -44,14 +44,14 @@
     scale: 0.25;
   }
   25%{
-    transform: translateX(15rem);
+    transform: translateY(10vh) translateX(15vw);
     scale: 0.65;
   }
   40%{
     opacity: 1;
   }
   50%{
-    transform: translateX(0rem);
+    transform: translateX(0);
     scale: 1;
   }
 }
