+++ src/components/Hero/Hero.jsx
@@ -11,7 +11,7 @@ function Hero () {
     </div>
     <h1 className="header header__three">Map your inner monolouge</h1>
     <section className="enterance">
-      <Link to={'/addcomment'} className="enterance__anonymous enterance__card">
+      <Link to={'/map'} className="enterance__anonymous enterance__card">
         <p>Post Anonymously</p>
         <p>Keep it local</p>
         <div className="ellpses">
